+++ src/index.html
@@ -1,14 +1,15 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Sleep Outside | Home</title>
-    <link rel="stylesheet" href="/css/style.css" />
-    <script src="../js/main.js" type="script"></script>
-    <script src="../js/product.js" type="script"></script>
+<head>
+  <meta charset="UTF-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Sleep Outside | Home</title>
+  <link rel="stylesheet" href="/css/style.css" />
+  <!--<script src="../js/main.js" type="module"></script>
+  <script src="../js/product.js" type="module"></script>-->
+  <!--<script src="../js/home.js" type="module"></script>-->
 
-  </head>
+</head>
   <body>
     <header class="divider">
       <div class="logo">
@@ -68,53 +69,49 @@
       </section>
       <section class="products">
         <h2>Top Products</h2>
-        <div id="index-product-display"></div>
-        <!--<ul class="product-list">
+        <!--<div id="index-product-display"></div>-->
+        <ul class="product-list">
     <li class="product-card">
-      <a href="product_pages/marmot-ajax-3.html">
+      <a href="product_pages/index.html?id=880RR">
         <img
           src="images/tents/marmot-ajax-tent-3-person-3-season-in-pale-pumpkin-terracotta~p~880rr_01~320.jpg"
           alt="Marmot Ajax tent"
         />
         <h3 class="card__brand">Marmot</h3>
         <h2 class="card__name">Ajax Tent - 3-Person, 3-Season</h2>
-        <p class="product-card__price">$199.99</p></a
-      >
+        <p class="product-card__price">$199.99</p></a>
     </li>
     <li class="product-card">
-      <a href="product_pages/northface-talus-4.html">
+      <a href="product_pages/index.html?id=985RF">
         <img
           src="images/tents/the-north-face-talus-tent-4-person-3-season-in-golden-oak-saffron-yellow~p~985rf_01~320.jpg"
           alt="Talus Tent - 4-Person, 3-Season"
         />
         <h3 class="card__brand">The North Face</h3>
         <h2 class="card__name">Talus Tent - 4-Person, 3-Season</h2>
-        <p class="product-card__price">$199.99</p></a
-      >
+        <p class="product-card__price">$199.99</p></a>
     </li>
     <li class="product-card">
-      <a href="product_pages/northface-alpine-3.html">
+      <a href="product_pages/index.html?id=985PR">
         <img
           src="images/tents/the-north-face-alpine-guide-tent-3-person-4-season-in-canary-yellow-high-rise-grey~p~985pr_01~320.jpg"
           alt="Alpine Guide Tent - 3-Person, 4-Season"
         />
         <h3 class="card__brand">The North Face</h3>
         <h2 class="card__name">Alpine Guide Tent - 3-Person, 4-Season</h2>
-        <p class="product-card__price">$349.99</p></a
-      >
+        <p class="product-card__price">$349.99</p></a>
     </li>
     <li class="product-card">
-      <a href="product_pages/cedar-ridge-rimrock-2.html">
+      <a href="product_pages/index.html?id=344YJ">
         <img
           src="images/tents/cedar-ridge-rimrock-tent-2-person-3-season-in-rust-clay~p~344yj_01~320.jpg"
           alt="Rimrock Tent - 2-Person, 3-Season"
         />
         <h3 class="card__brand">Cedar Ridge</h3>
         <h2 class="card__name">Rimrock Tent - 2-Person, 3-Season</h2>
-        <p class="product-card__price">$69.99</p></a
-      >
+        <p class="product-card__price">$69.99</p></a>
     </li>
-  </ul>-->
+  </ul>
       </section>
     </main>
     <footer>&copy;NOT a real business</footer>
